+++ src/matchers/matcher.rs
@@ -2,6 +2,6 @@ pub trait Matcher <T> {
     fn assert_check(&self, expected: T) -> bool;
     fn msg(&self, expected: T) -> String;
     fn negated_msg(&self, expected: T) -> String;
-    fn get_file_line(&self) -> (&'static str, usize);
+    fn get_file_line(&self) -> (&'static str, u32);
 }
